Fixed bug in determinant function and results can be obtained up to long long int values

+++ matrix_operations/operations.h
@@ -2,6 +2,6 @@
 #define OPERATIONS_H
 
 void show(std::vector<std::vector< int>>& tab);
-int determinant( std::vector<std::vector<int>>& tab);
+long long int determinant( std::vector<std::vector<int>>& tab);
 
 #endif
